+++ NEWS.md
@@ -11,6 +11,7 @@
 - Changed some arguments to strict matching (`rlang::arg_match()`) instead of partial matching (`match.arg()`) (e.g. `mut_type = "s"` doesn't work anymore and must be fully specified `mut_type = "somatic_only"`).
 - Added unit tests for gnomeR plots/visuals (#144).
 - A dictionary of old to new names for `rename_columns()` output is now an attribute of the returned object. Now messages can reference the original names of data columns (ex: `TumorAllele2` not `tumor_allele_2`) to make it more intuitive to users (#302).
+- Fixed bug that wasn't consistently filtering out germline samples
 
 
 # gnomeR 1.2.0